created separate app.js file
