+++ server.js
@@ -1,24 +1,10 @@
-const express = require('express');
-const cors = require('cors');
 const mongoose = require('mongoose');
 const config = require('./config');
 const path = require('path');
+const app = require('./app');
 
-const users = require('./routes/users');
-const auth = require('./routes/auth');
-
-const { MONGO_URI } = config;
-
-const app = express();
 const port = process.env.PORT || 5000;
-
-// middleware
-app.use(cors());
-app.use(express.json());
-
-// routes 
-app.use('/api/users', users);
-app.use('/api/auth', auth);
+const { MONGO_URI } = config;
 
 // connect to database
 mongoose
